bd_service/requirements.txt
alembic>=1.13
SQLAlchemy>=2.0
python-dotenv>=1.0
